chore: helper class
