+++ packages/fx-core/src/component/resource/spfx/utils/questions.ts
@@ -10,12 +10,14 @@ import {
 } from "../error";
 import { Constants } from "./constants";
 import { Utils } from "./utils";
+import { PackageSelectOptionsHelper } from "./question-helper";
 
 export enum SPFXQuestionNames {
   framework_type = "spfx-framework-type",
   webpart_name = "spfx-webpart-name",
   webpart_desp = "spfx-webpart-desp",
   version_check = "spfx-version-check",
+  load_package_version = "spfx-load-package-version",
   use_global_package_or_install_local = "spfx-use-global-package-or-install-local",
 }
 
@@ -106,14 +108,12 @@ export const versionCheckQuestion: Question = {
   },
 };
 
-let options: OptionItem[] = [];
-
-export const loadDataQuestion: Question = {
+export const loadPackageVersions: Question = {
   type: "func",
-  name: "loadQuestion",
+  name: SPFXQuestionNames.load_package_version,
   title: getLocalizedString("plugins.spfx.questions.packageSelect.title"),
   func: async (inputs: Inputs) => {
-    await getDynamicOptions();
+    await PackageSelectOptionsHelper.loadOptions();
     return undefined;
   },
 };
@@ -122,55 +122,9 @@ export const spfxPackageSelectQuestion: Question = {
   type: "singleSelect",
   name: SPFXQuestionNames.use_global_package_or_install_local,
   title: getLocalizedString("plugins.spfx.questions.packageSelect.title"),
-  staticOptions: options,
-  // [
-  // {
-  //   id: "globalPackage",
-  //   label: getLocalizedString("plugins.spfx.questions.packageSelect.useGlobalPackage.label"),  // todo: get version
-  //   detail: getLocalizedString("plugins.spfx.questions.packageSelect.useGlobalPackage.detail", Constants.RecommenLowestSpfxVersion)
-  // },
-  // { id: "installLocally",
-  // label: getLocalizedString("plugins.spfx.questions.packageSelect.installLocally.label")}, // todo: get version
-  //],
+  staticOptions: [],
   placeholder: getLocalizedString("plugins.spfx.questions.packageSelect.placeholder"),
   dynamicOptions: async (inputs: Inputs): Promise<OptionItem[]> => {
-    return options;
+    return PackageSelectOptionsHelper.getOptions();
   },
 };
-
-async function getDynamicOptions(): Promise<OptionItem[]> {
-  let versions: (string | undefined)[] = [undefined, undefined];
-
-  versions = await Promise.all([
-    Utils.findGloballyInstalledVersion(undefined, Constants.GeneratorPackageName, 5, false),
-    Utils.findLatestVersion(undefined, Constants.GeneratorPackageName, 10),
-  ]);
-
-  const result: OptionItem[] = [
-    {
-      id: "globalPackage",
-      label:
-        versions[0] !== undefined
-          ? getLocalizedString(
-              "plugins.spfx.questions.packageSelect.useGlobalPackage.withVersion.label",
-              versions[0]
-            )
-          : getLocalizedString(
-              "plugins.spfx.questions.packageSelect.useGlobalPackage.noVersion.label"
-            ),
-      detail: getLocalizedString(
-        "plugins.spfx.questions.packageSelect.useGlobalPackage.detail",
-        Constants.RecommenLowestSpfxVersion
-      ),
-    },
-    {
-      id: "installLocally",
-      label: getLocalizedString(
-        "plugins.spfx.questions.packageSelect.installLocally.label",
-        versions[1]
-      ),
-    },
-  ];
-  options = result;
-  return result;
-}
